move network map client to node
